// src/app/models/common/databaseresult.ts
export class DatabaseResult {

    idResponseCode: number;

    responseCode: string;
    
    idactiontransaction: number;
    
    table: string;

    objectname: string;

    attributename: string;

    entityid: number;

    errorId: number;

    message: string;
    
    userid: number;
            
}